Add support for gzipped files
When a file is compressed, it is supposed to carry the `Content-Type` of the underlying compressed file
type, e.g. `data.js.gz` should be served as `application/javascript`. In these cases, `Content-Encoding`
should also be set to `gzip`.

This could be extended in future to support other encodings (`compress`, `deflate`, `br`).